generated_code/fixtures/tests/test_dynamicexampleprompt_fixture.py
import pytest
from generated_code.fixtures.fixtures.dynamicexampleprompt_fixture import fixture_mock_dynamicexampleprompt
from code_autoeval.llm_model.utils.prompting.dynamic_example_prompt import DynamicExamplePrompt
def test_mock_dynamicexampleprompt(fixture_mock_dynamicexampleprompt):
    assert isinstance(fixture_mock_dynamicexampleprompt, DynamicExamplePrompt)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_act')
    assert callable(fixture_mock_dynamicexampleprompt.generate_act)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_arrange')
    assert callable(fixture_mock_dynamicexampleprompt.generate_arrange)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_assert')
    assert callable(fixture_mock_dynamicexampleprompt.generate_assert)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_fixture_check')
    assert callable(fixture_mock_dynamicexampleprompt.generate_fixture_check)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_init_test')
    assert callable(fixture_mock_dynamicexampleprompt.generate_init_test)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'generate_mock_return_value')
    assert callable(fixture_mock_dynamicexampleprompt.generate_mock_return_value)
    assert hasattr(fixture_mock_dynamicexampleprompt, 'provide_test_structure')
    assert callable(fixture_mock_dynamicexampleprompt.provide_test_structure)